Add missing intermediate certificate in cert loader cache
