+++ components/director/pkg/certloader/loader.go
@@ -8,6 +8,8 @@ import (
 	"encoding/pem"
 	"time"
 
+	"github.com/kyma-incubator/compass/components/director/pkg/cert"
+
 	"github.com/kyma-incubator/compass/components/director/pkg/kubernetes"
 	"github.com/kyma-incubator/compass/components/director/pkg/namespacedname"
 
@@ -142,21 +144,16 @@ func (cl *certificatesLoader) processEvents(ctx context.Context, events <-chan w
 
 func (cl *certificatesLoader) parseCertificate(ctx context.Context, secretData map[string][]byte) (*tls.Certificate, error) {
 	log.C(ctx).Info("Parsing certificate data from secret...")
-	certBytes := secretData["tls.crt"]
+	certChainBytes := secretData["tls.crt"]
 	privateKeyBytes := secretData["tls.key"]
 
-	if certBytes == nil || privateKeyBytes == nil {
+	if certChainBytes == nil || privateKeyBytes == nil {
 		return nil, errors.New("There is no certificate data in the secret")
 	}
 
-	clientCrtPem, _ := pem.Decode(certBytes)
-	if clientCrtPem == nil {
-		return nil, errors.New("Error while decoding certificate pem block")
-	}
-
-	clientCert, err := x509.ParseCertificate(clientCrtPem.Bytes)
+	certs, err := cert.DecodeCertificates(certChainBytes)
 	if err != nil {
-		return nil, err
+		return nil, errors.New("Error while decoding certificate pem block")
 	}
 
 	privateKeyPem, _ := pem.Decode(privateKeyBytes)
@@ -178,8 +175,7 @@ func (cl *certificatesLoader) parseCertificate(ctx context.Context, secretData m
 	}
 
 	log.C(ctx).Info("Successfully parse certificate from secret data")
-	return &tls.Certificate{
-		Certificate: [][]byte{clientCert.Raw},
-		PrivateKey:  privateKey,
-	}, nil
+	tlsCert := cert.NewTLSCertificate(privateKey, certs...)
+
+	return &tlsCert, nil
 }
